pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eks-cost-estimator"
version = "0.1.0"
description = "CLI to estimate EKS costs from Kubernetes YAML manifests before deployment"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "EKS Cost Estimator" }
]
dependencies = [
  "typer>=0.9.0",
  "rich>=13.6.0",
  "PyYAML>=6.0",
  "pydantic>=2.3",
]

[project.optional-dependencies]
aws = [
  "boto3>=1.28.0",
]

[project.scripts]
eks-cost-estimator = "eks_cost_estimator.cli.main:app"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203"]
per-file-ignores = [
  "tests/*: S101",
]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
ignore_missing_imports = true
plugins = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["eks_cost_estimator*"]
